Combine get vacation details and comments UseCases into GetVacationDataUseCase
