+++ app/src/main/java/com/planzy/app/domain/usecase/vacation/GetVacationDataUseCase.kt
@@ -2,6 +2,7 @@ package com.planzy.app.domain.usecase.vacation
 
 import com.planzy.app.domain.model.Place
 import com.planzy.app.domain.model.Vacation
+import com.planzy.app.domain.model.VacationComment
 import com.planzy.app.domain.repository.PlacesRepository
 import com.planzy.app.domain.repository.VacationsRepository
 
@@ -11,11 +12,11 @@ data class VacationDetails(
     val places: List<Place>
 )
 
-class GetVacationDetailsUseCase(
+class GetVacationDataUseCase(
     private val vacationsRepository: VacationsRepository,
     private val placesRepository: PlacesRepository
 ) {
-    suspend operator fun invoke(vacationId: String): Result<VacationDetails> {
+    suspend fun getVacationDetails(vacationId: String): Result<VacationDetails> {
         return try {
             val vacationResult = vacationsRepository.getVacationWithUser(vacationId)
 
@@ -25,13 +26,13 @@ class GetVacationDetailsUseCase(
 
                     placeIdsResult.fold(
                         onSuccess = { placeIds ->
-                             val places = placeIds.mapIndexedNotNull { index, placeId ->
+                             val places = placeIds.mapIndexedNotNull { _, placeId ->
                                 val placeResult = placesRepository.getPlaceDetails(placeId)
                                 placeResult.fold(
                                     onSuccess = { place ->
                                         place
                                     },
-                                    onFailure = { error ->
+                                    onFailure = { _ ->
                                         null
                                     }
                                 )
@@ -57,5 +58,9 @@ class GetVacationDetailsUseCase(
         } catch (e: Exception) {
             Result.failure(e)
         }
+    }
+
+    suspend fun getVacationComments(vacationId: String): Result<List<VacationComment>> {
+        return vacationsRepository.getVacationComments(vacationId)
     }
 }
